# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OutputFormat = (typeof OutputFormat)[keyof typeof OutputFormat];

export const OutputFormat = {
  docx: "docx",
  xlsx: "xlsx",
  pptx: "pptx",
  png: "png",
  jpg: "jpg",
  txt: "txt",
} as const;

export type ConversionStatus =
  (typeof ConversionStatus)[keyof typeof ConversionStatus];

export const ConversionStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface ConversionRecord {
  id: number;
  fileId: number;
  outputFormat: OutputFormat;
  status: ConversionStatus;
  convertedPath?: string | null;
  errorMessage?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface FileRecord {
  id: number;
  originalName: string;
  storedPath: string;
  sizeBytes: number;
  uploadedAt: string;
  conversions?: ConversionRecord[];
}

export interface ConversionRequest {
  fileId: number;
  outputFormat: OutputFormat;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListFiles200 = {
  files: FileRecord[];
};

export type UploadFileBody = {
  file: Blob;
};

export type DeleteFile200 = {
  success: boolean;
};