# lib/api-zod/src/generated/types/conversionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionStatus } from "./conversionStatus";
import type { OutputFormat } from "./outputFormat";

export interface ConversionRecord {
  id: number;
  fileId: number;
  outputFormat: OutputFormat;
  status: ConversionStatus;
  convertedPath?: string | null;
  errorMessage?: string | null;
  createdAt: Date;
  completedAt?: Date | null;
}