# lib/api-zod/src/generated/types/conversionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OutputFormat } from "./outputFormat";

export interface ConversionRequest {
  fileId: number;
  outputFormat: OutputFormat;
}