# lib/api-zod/src/generated/types/conversionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */

export type ConversionStatus =
  (typeof ConversionStatus)[keyof typeof ConversionStatus];

export const ConversionStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;