# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionRecord } from "./conversionRecord";

export interface FileRecord {
  id: number;
  originalName: string;
  storedPath: string;
  sizeBytes: number;
  uploadedAt: Date;
  conversions?: ConversionRecord[];
}