# lib/api-zod/src/generated/types/outputFormat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PDF Management & Conversion Platform API
 * OpenAPI spec version: 0.1.0
 */

export type OutputFormat = (typeof OutputFormat)[keyof typeof OutputFormat];

export const OutputFormat = {
  docx: "docx",
  xlsx: "xlsx",
  pptx: "pptx",
  png: "png",
  jpg: "jpg",
  txt: "txt",
} as const;